Use shared sources for macOS and iOS
Flutter >= 3.7.0 supports sharing source code between iOS and macOS
plugins. The iOS code in `quick_blue` was removed and replaced with just
the macOS federated plugin, which was modified to support both iOS and
macOS.
