--- quick_blue/ios/Classes/QuickBluePlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface QuickBluePlugin : NSObject<FlutterPlugin>
-@end
