--- quick_blue/ios/Classes/QuickBluePlugin.m
@@ -1,15 +1,0 @@
-#import "QuickBluePlugin.h"
-#if __has_include(<quick_blue/quick_blue-Swift.h>)
-#import <quick_blue/quick_blue-Swift.h>
-#else
-// Support project import fallback if the generated compatibility header
-// is not copied when this plugin is created as a library.
-// https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
-#import "quick_blue-Swift.h"
-#endif
-
-@implementation QuickBluePlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftQuickBluePlugin registerWithRegistrar:registrar];
-}
-@end
