--- quick_blue/macos/.gitignore
@@ -1,6 +1,0 @@
-# Flutter-related
-**/Flutter/ephemeral/
-**/Pods/
-
-# Xcode-related
-**/xcuserdata/
