+++ quick_blue_example/pubspec.yaml
@@ -1,5 +1,6 @@
 name: quick_blue_example
 description: Demonstrates how to use the quick_blue plugin.
+version: 0.1.0+1
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
