+++ quick_blue_macos/darwin/.gitignore
