+++ quick_blue_macos/darwin/quick_blue_macos.podspec
@@ -1,9 +1,9 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint quick_blue.podspec' to validate before publishing.
+# Run `pod lib lint quick_blue_macos.podspec' to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'quick_blue'
+  s.name             = 'quick_blue_macos'
   s.version          = '0.0.1'
   s.summary          = 'A new flutter plugin project.'
   s.description      = <<-DESC
@@ -13,8 +13,13 @@ A new flutter plugin project.
   s.license          = { :file => '../LICENSE' }
   s.author           = { 'Your Company' => 'email@example.com' }
   s.source           = { :path => '.' }
-  s.source_files = 'Classes/**/*'
+  s.source_files     = 'Classes/**/*'
+  s.osx.dependency 'FlutterMacOS'
+  s.ios.dependency 'Flutter'
 
-  s.platform = :osx
-  s.osx.deployment_target = '10.11'
+  s.platform = :osx, '10.14'
+  s.ios.deployment_target = '11.0'
+  s.osx.deployment_target = '10.14'
+  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
+  s.swift_version = '5.0'
 end
